+++ src/reanimated2/NativeReanimated/index.web.ts
@@ -1,0 +1,4 @@
+'use strict';
+import reanimatedJS from '../js-reanimated';
+
+export default reanimatedJS;
