+++ src/reanimated2/SetNativeProps.ts
@@ -1,15 +1,9 @@
 'use strict';
 import type { ShadowNodeWrapper, StyleProps } from './commonTypes';
-import {
-  isChromeDebugger,
-  isJest,
-  isWeb,
-  shouldBeUseWeb,
-} from './PlatformChecker';
+import { isChromeDebugger, isJest, shouldBeUseWeb } from './PlatformChecker';
 
 import type { AnimatedRef } from './hook/commonTypes';
 import type { Component } from 'react';
-import { _updatePropsJS } from './js-reanimated';
 import { processColorsInProps } from './Colors';
 
 const IS_NATIVE = !shouldBeUseWeb();
@@ -19,12 +13,7 @@ export let setNativeProps: <T extends Component>(
   updates: StyleProps
 ) => void;
 
-if (isWeb()) {
-  setNativeProps = (_animatedRef, _updates) => {
-    const component = (_animatedRef as any)();
-    _updatePropsJS(_updates, { _component: component });
-  };
-} else if (IS_NATIVE && global._IS_FABRIC) {
+if (IS_NATIVE && global._IS_FABRIC) {
   setNativeProps = (animatedRef, updates) => {
     'worklet';
     if (!_WORKLET) {
@@ -53,15 +42,15 @@ if (isWeb()) {
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     _updatePropsPaper!([{ tag, name, updates }]);
   };
+} else if (isJest()) {
+  setNativeProps = () => {
+    console.warn('[Reanimated] setNativeProps() is not supported with Jest.');
+  };
 } else if (isChromeDebugger()) {
   setNativeProps = () => {
     console.warn(
       '[Reanimated] setNativeProps() is not supported with Chrome Debugger.'
     );
-  };
-} else if (isJest()) {
-  setNativeProps = () => {
-    console.warn('[Reanimated] setNativeProps() is not supported with Jest.');
   };
 } else {
   setNativeProps = () => {
