+++ src/reanimated2/setGestureState.web.ts
@@ -1,0 +1,8 @@
+'use strict';
+
+export const setGestureState: (
+  handlerTag: number,
+  newState: number
+) => void = () => {
+  console.warn('[Reanimated] setGestureState() is not available on web.');
+};
