+++ src/reanimated2/utils.ts
@@ -1,6 +1,6 @@
 'use strict';
 import type { Component } from 'react';
-import { measure } from './NativeMethods';
+import { measure } from './measure';
 import type { AnimatedRef } from './hook/commonTypes';
 import type { SharedValue } from './commonTypes';
